+++ .gitignore
@@ -1,0 +1,8 @@
+*.pyc
+.env
+.DS_Store
+db.sqlite3
+media/
+venv/
+*/migrations/*
+!*/migrations/__init__.py
