--- dashboard/templates/base.html
@@ -1,52 +1,0 @@
-{%load static %}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Bootstrap 5 CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
-    <!-- Font awesome-->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css" integrity="sha512-1ycn6IcaQQ40/MKBW2W4Rhis/DbILU74C1vSrLJxCq57o941Ym01SwNsOMqvEBFlcgUa6xLiPY/NS5R+E6ztJQ==" crossorigin="anonymous" referrerpolicy="no-referrer" />
-    <title>{% block title%} {% endblock title %}</title>
-     <!-- Favicon -->
-    <link href={% static "img/favicon.ico" %} rel="shortcut icon">
-    <!-- Logo -->
-    <link href={% static "img/logo.ico" %} rel="shortcut icon">
-    <!-- Static CSS-->
-    <link href={% static "css/style.css" %} rel="stylesheet">
-</head>
-<body class="bg-dark">
-    <div class="wrapper">
-        <!-- Navbar -->
-        <header>
-            <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-                <div class="col-sm">
-                    <a class="navbar-brand flex-grow-1" style="text-align:center; color:azure; margin-left: 25px;">
-                        <i class="fas fa-document"></i> &nbsp;&nbsp; Logo
-                    </a>
-                </div>
-                <div>
-                    {%  block nav-btn %}
-                        <a href="{% url 'data_read' %}" id="btn-nav" class="btn btn-outline-light">Django Hub</a>
-                    {%  endblock nav-btn %}
-                </div>
-            </nav>
-        </header>
-        {% block content%} {% endblock content %}
-    
-    </div> <!-- END // WRAPPER -->
-
-    <footer class="bg-dark text-center text-lg-start" id="footer" style="color:azure">
-        <div class="text-center p3" style="background-color: rgba(0,0,0,0.2);">
-            &copy; {% now "Y" %} Copyright:
-            <a class="text-align" style="text-decoration: none;">
-                <span class="text-warning"> www.github.com/stnbnvdz </span>
-            </a>
-        </div>
-    </footer>
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous"></script>
-</body>
-</html>
