--- dashboard/templates/data.read.html
@@ -1,48 +1,0 @@
-{% extends 'base.html' %}
-
-{% block nav-btn %}
-<a href="{% url 'data_Create' %}" id="btn-nav" class="btn btn-outline-light">Go Home</a>
-{% endblock nav-btn %}
-
-{% block content%}
-<div class="container mt-5">
-    <h5 class="card p-2 bg-warning text-center mb-3">List of Resources</h5>
-    <table class="table table-bordered text-light" style="text-align:center">
-        <thead class="table-warning">
-            <tr>
-                <th style="width: 20%;">Name</th>
-                <th style="width: 30%;">Link</th>
-                <th style="width: 10%;">Tags</th>
-                <th style="width: 15%;">Format</th>
-                <th style="width: 15%;">Difficulty</th>
-                <th style="width: 20%;">Est. Time of Completion</th>
-                <th colspan="2" style="width: 10%;">Action</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for data in data_read %}
-            <tr>
-                <td>{{ data.name }}</td>
-                <td>{{ data.link }}</td>
-                <td>{{ data.tags }}</td>
-                <td>{{ data.difficulty }}</td>
-                <td>{{ data.time }}</td>
-                <td class="bg-warning">
-                    <a href="{% url 'data_update' data.id %}" class="btn btn-sm">
-                        <i class="fas fa-pen f-size"></i>
-                    </a>
-                </td>
-                <td class="bg-secondary">
-                    <a href="/data_delete/{{ data. id }}" class="btn btn-sm text-light">
-                        <i class="fas fa-trash-alt f-size"></i>
-                    </a>
-                </td>
-                {% empty %}
-                <td colspan="6" style="color:firebrick">No records found</td>
-            </tr>
-            {% endfor %}
-        </tbody>
-    </table>
-</div>
-
-{% endlock content%}
