--- dashboard/templates/data_form.html
@@ -1,28 +1,0 @@
-{% extends 'base.html' %}
-
-<!-- Crispy Forms -->
-{% load crispy_forms_tags %}
-
-{% block content %}
-<div class="container">
-    <h5 class="text-center">
-        <strong><i class="fas fa-plus"></i>&nbsp;>&nbsp; Add Resources</strong>
-    </h5>
-    <hr>
-    <form method="POST" autocomplete="off">
-        {% csrf_token %}
-        {{ form.name|as_crispy_field }}
-        {{ form.link|as_crispy_field }}
-        {{ form.tags|as_crispy_field }}
-        {{ form.format|as_crispy_field }}
-        {{ form.difficulty|as_crispy_field }}
-        {{ form.time|as_crispy_field }}
-
-        <div class="mt-5">
-            <hr>
-            <button type="submit" class="btn btn-warning btn-width">Submit</button>
-            <button type="submit" class="btn btn-secondary ms-2 btn-width">Reset</button>
-        </div>
-    </form>
-</div>
-{% endblock content %}
