--- django_crud/urls.py
@@ -1,11 +1,0 @@
-from django.contrib import admin
-from django.urls import path
-from dashboard import views
-
-urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', views.data_form, name='data_create'),
-    path('data/', views.data_read, name='data_read'),
-    path('<int:id>', views.data_form, name='data_update'),
-    path('data_delete/<str:candidate_id>', views.data_delete, name='data_delete'),
-]
