+++ requirements.txt
@@ -1,0 +1,3 @@
+asgiref==3.5.2
+Django==4.0.4
+sqlparse==0.4.2
