+++ resources/templates/resources/add.html
@@ -1,0 +1,29 @@
+{% extends "resources/base.html" %}
+
+{% block body %}
+  <h3 class="text-center m-4">Add Resources</h3>
+  {% if success %}
+    <div class="alert alert-success" role="alert">
+      New resources added successfully.
+      <a href="{% url 'index' %}" class="alert-link">Go to Home Page.</a>
+    </div>
+  {% else %}
+    <div class="row justify-content-center">
+      <div class="col-6">
+        <div class="card bg-light mb-3">
+          <div class="card-header">
+            <i class="fa-solid fa-address-card fa-lg"></i> Resources
+          </div>
+          <div class="card-body">
+            <form action="{% url 'add' %}" method="POST">
+              {% csrf_token %}
+              {{ form.as_p }}
+              <button type="submit" class="btn btn-primary">Add</button>
+              <a href="{% url 'index' %}" class="btn btn-secondary">Cancel</a>
+            </form>
+          </div>
+        </div>
+      </div>
+    </div>
+  {% endif %}
+{% endblock %}
