+++ resources/templates/resources/index.html
@@ -1,0 +1,117 @@
+{% extends "resources/base.html" %}
+
+{% block body %}
+<h3 class="m-4">All Resources</h3>
+<div class="row">
+  <div class="col-12">
+    {% if resources %}
+    <div class="card bg-light ms-4 me-4 mb-4">
+      <div class="card-header">
+        <i class="fa-solid fa-list fa-lg"></i> Resources List
+      </div>
+      <div class="card-body">
+        <p class="card-text">
+          <div class="table-responsive">
+            <table class="table table-hover">
+              <thead>
+                <tr>
+                  <th scope="col">Name</th>
+                  <th scope="col">Link</th>
+                  <th scope="col">Tags</th>
+                  <th scope="col">Format</th>
+                  <th scope="col">Difficulty</th>
+                  <th scope="col">Estimated Time of Completion</th>
+                  <th scope="col">Actions</th>
+                </tr>
+              </thead>
+              <tbody>
+                {% for resources in resources %}
+                  <tr>
+                    <td>{{ resources.name }}</td>
+                    <td>{{ resources.link }}</td>
+                    <td>{{ resources.tags }}</td>
+                    <td>{{ resources.format }}</td>
+                    <td>{{ resources.difficulty }}</td>
+                    <td>{{ resources.time }}</td>
+                    <td>
+                      <button type="button" class="btn btn-success me-1" href="{% url 'view_resources' resources.id %}" data-bs-toggle="modal" data-bs-target="#myModal{{resources.id}}">
+                        <i class="fa-solid fa-circle-info fa-lg"></i>
+                      </button>
+
+                      <!-- Modal - View  -->
+                      <div class="modal" id="myModal{{resources.id}}" tabindex="-1" aria-labelledby="myModalLabel1" aria-hidden="true">
+                        <div class="modal-dialog" role="document">
+                          <div class="modal-content">
+                            <div class="modal-header">
+                              <h5 class="modal-title">
+                                Resources: {{ resources.name }}
+                              </h5>
+                              <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close">
+                                <span aria-hidden="true"></span>
+                              </button>
+                            </div>
+                            <div class="modal-body">
+                              <ul class="list-unstyled">
+                                <li>Name: <strong>{{ resources.name }}</strong></li>
+                                <li>Link: <strong>{{ resources.link }}</strong></li>
+                                <li>Tags: <strong>{{ resources.tags }}</strong></li>
+                                <li>Format: <strong>{{resources.format }}</strong></li>
+                                <li>Difficulty: <strong>{{ resources.difficulty }}</strong></li>
+                                <li>Time: <strong>{{ resources.time }}</strong></li>
+                              </ul>
+                            </div>
+                            <div class="modal-footer">
+                              <button type="button" class="btn btn-primary" data-bs-dismiss="modal">Close</button>
+                            </div>
+                          </div>
+                        </div>
+                      </div>
+
+                      <a class="btn btn-warning me-1" href="{% url 'edit' resources.id %}">
+                        <i class="fa-solid fa-pen-to-square fa-lg"></i>
+                      </a>
+
+                      <button type="button" class="btn btn-danger" data-bs-toggle="modal" data-bs-target="#delete{{resources.id}}">
+                        <i class="fa-solid fa-trash-can fa-lg"></i>
+                      </button>
+
+                      <!-- Modal - Delete -->
+                      <div class="modal" id="delete{{resources.id}}" tabindex="-1" aria-labelledby="myModalLabel2" aria-hidden="true">
+                        <div class="modal-dialog" role="document">
+                          <div class="modal-content">
+                            <div class="modal-header">
+                              <h5 class="modal-title">
+                                Name: {{ resources.name }}
+                              </h5>
+                              <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close">
+                                <span aria-hidden="true"></span>
+                              </button>
+                            </div>
+                            <div class="modal-body">
+                              <p>Are you sure you want?</p>
+                            </div>
+                            <div class="modal-footer">
+                              <form action="{% url 'delete' resources.id %}" method="POST">
+                                {% csrf_token %}
+                                <input type="submit" class="btn btn-primary" value="Delete">
+                              </form>
+                              <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
+                            </div>
+                          </div>
+                        </div>
+                      </div>
+                    </td>
+                  </tr>
+                {% endfor %}
+              </tbody>
+            </table>
+          </div>
+        </p>
+      </div>
+    </div>
+    {% else %}
+      <h5 class="alert alert-primary ms-4 me-4">No Records</h5>
+    {% endif %}
+  </div>
+</div>
+{% endblock %}
